third_party/WebKit/LayoutTests/fast/events/inputevents/before-input-ranges.html

<!DOCTYPE html>
<html>
<head>
<title>InputEvent: beforeinput getTargetRanges()</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<p id="sample" contentEditable="true"></p>
<textarea id="textarea1"></textarea>
<div id="log"></div>
<script>
test(function() {
    if (!window.eventSender) {
        document.write('This test requires eventSender');
        return;
    }

    // Setup data.
    var sample = document.getElementById('sample');
    sample.innerHTML = 'hel<i>lo wo</i><b>rld</b>';

    // Move caret right after 'hello world '.
    var selection = window.getSelection();
    selection.collapse(sample.firstChild, 0);
    selection.modify('move', 'forward', 'line');

    // Check event for delete-word-backward.
    var handler = function(event) {
        var ranges = event.getTargetRanges();
        assert_equals(ranges.length, 1);
        var range = ranges[0];
        assert_true(range instanceof StaticRange);
        assert_equals(range.startOffset, 3);
        assert_equals(range.startContainer.textContent, 'lo wo');
        assert_equals(range.endOffset, 3);
        assert_equals(range.endContainer.textContent, 'rld');
        sample.removeEventListener('beforeinput', handler);
    };
    sample.addEventListener('beforeinput', handler);

    // Send delete-word-backward, should delete 'world' across I and B.
    var isMacOSX = navigator.userAgent.indexOf("Mac OS X") != -1;
    eventSender.keyDown('Backspace', [isMacOSX ? 'altKey' : 'ctrlKey']);
    assert_equals(sample.innerHTML, 'hel<i>lo&nbsp;</i>');
}, 'Testing getTargetRanges() returns correct range across elements.');

test(function() {
    if (!window.eventSender) {
        document.write('This test requires eventSender');
        return;
    }

    // Setup data.
    var sample = document.getElementById('sample');
    sample.innerHTML = 'hello';

    // Move caret right after 'hello'.
    var selection = window.getSelection();
    selection.collapse(sample.firstChild, 0);
    selection.modify('move', 'forward', 'word');

    // Check and hold event for delete-character-backward.
    var lastEvent;
    var handler = function(event) {
        lastEvent = event;
        assert_equals(lastEvent.getTargetRanges().length, 1);
        sample.removeEventListener('beforeinput', handler);
    };
    sample.addEventListener('beforeinput', handler);

    // Send delete-character-backward.
    eventSender.keyDown('Backspace', []);
    assert_equals(sample.textContent, 'hell');
    // |getRanges| should be cleared after dispatch.
    assert_equals(lastEvent.getTargetRanges().length, 0);
}, 'Testing getTargetRanges() cleared after dispatch.');

test(function() {
    assert_not_equals(window.eventSender, undefined, 'This test requires eventSender.');
    assert_not_equals(window.testRunner, undefined, 'This test requires testRunner.');

    // Setup data.
    const sample = document.getElementById('sample');
    sample.innerHTML = 'hello';

    // Move caret right after 'hello'.
    const selection = window.getSelection();
    selection.collapse(sample.firstChild, 0);
    selection.modify('move', 'forward', 'word');

    // Record target ranges.
    let lastRangeString = '';
    const handler = function(event) {
        const range = event.getTargetRanges()[0];
        lastRangeString = `${range.startContainer.textContent}-${range.startOffset}-${range.endContainer.textContent}-${range.endOffset}`;
    };
    sample.addEventListener('beforeinput', handler);

    // Type 'a' at the end of text.
    eventSender.keyDown('a');
    assert_equals(sample.textContent, 'helloa');
    assert_equals(lastRangeString, 'hello-5-hello-5');

    // Select all text and type 'b'.
    selection.modify('extend', 'backward', 'word');
    eventSender.keyDown('b');
    assert_equals(sample.textContent, 'b');
    assert_equals(lastRangeString, 'helloa-0-helloa-6');

    // Select all text and bold.
    selection.modify('extend', 'backward', 'word');
    testRunner.execCommand('bold');
    assert_equals(sample.textContent, 'b');
    assert_equals(lastRangeString, 'b-0-b-1');

    sample.removeEventListener('beforeinput', handler);
}, 'Actions other than deletion should have current selection as target ranges.');

test(function() {
    assert_not_equals(window.eventSender, undefined, 'This test requires eventSender.');

    // Setup data.
    const textarea1 = document.getElementById('textarea1');
    textarea1.value = 'hello';
    textarea1.focus();

    // Record range count.
    let lastRangeCount = -1;
    const handler = function(event) {
        lastRangeCount = event.getTargetRanges().length;
    };
    textarea1.addEventListener('beforeinput', handler);

    // Type 'a' at the end of text.
    lastRangeCount = -1;
    eventSender.keyDown('a');
    assert_equals(textarea1.value, 'helloa');
    assert_equals(lastRangeCount, 0);

    // Press backspace.
    lastRangeCount = -1;
    eventSender.keyDown('Backspace');
    assert_equals(textarea1.value, 'hello');
    assert_equals(lastRangeCount, 0);

    textarea1.removeEventListener('beforeinput', handler);
}, 'Textarea should have empty target range.');
</script>
</body>
</html>
